Extract AcPhoneNumberField in common form library
